ONLYHALFSTRATEGY/python/requirements.txt
grpcio==1.18.0
numpy==1.16.1
protobuf==3.6.1
six==1.12.0
TA_Lib-0.4.18-cp37-cp37m-win32.whl
